pyproject.toml
[tool.poetry]
name = "area"
version = "0.1.0"
description = ""
authors = ["haf <haf@incubtek.com>"]
packages = [{ include = "scripts"}]

[tool.poetry.dependencies]
python = "3.12.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
pyyaml = "^6.0.2"
typer = "^0.12.5"

[tool.poetry.scripts]
generate-data = 'scripts.main:app'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
